requirements.txt
flask
flask_mysqldb
flask_wtf
pyjwt